CMS-1935: Extract season helper functions from seasons route
