+++ backend/routes/api/seasons.js
@@ -4,9 +4,7 @@ import asyncHandler from "express-async-handler";
 import { Op } from "sequelize";
 import sequelize from "../../db/connection.js";
 import * as STATUS from "../../constants/seasonStatus.js";
-import * as DATE_TYPE from "../../constants/dateType.js";
 import * as SEASON_TYPE from "../../constants/seasonType.js";
-import * as FEATURE_TYPE from "../../constants/featureType.js";
 import {
   getAllDateTypes,
   getDateTypesForFeature,
@@ -18,16 +16,9 @@ import { resolveSeasonApprovalState } from "../../utils/seasonApprovalHelpers.js
 import {
   Park,
   Season,
-  FeatureType,
   Feature,
-  DateType,
-  DateRange,
-  DateRangeAnnual,
-  Dateable,
-  GateDetail,
   ParkArea,
   SeasonChangeLog,
-  DateChangeLog,
   User,
 } from "../../models/index.js";
 
@@ -41,666 +32,26 @@ import checkUserRoles, {
   getRolesFromAuth,
 } from "../../utils/checkUserRoles.js";
 
+import {
+  changeLogsQueryPart,
+  dateableAndDatesQueryPart,
+  featureTypeQueryPart,
+  SEASON_ATTRIBUTES,
+} from "../../utils/seasonQueryHelpers.js";
+import {
+  getPreviousSeasonDates,
+  getDateRangeAnnuals,
+  getGateDetail,
+  getFrontcountryFeatureReservationDates,
+  getParkDates,
+  getWinterSeason,
+} from "../../utils/seasonDataHelpers.js";
+import {
+  checkSeasonExists,
+  saveSeasonData,
+} from "../../utils/saveSeasonData.js";
+
 const router = Router();
-
-/**
- * Checks if a Season exists, and throws an error if the Season is not found.
- * @param {Season} season The Season model instance to check
- * @throws {Error} If the Season is not found, an error with status 404 is thrown
- * @returns {boolean} Returns true if the Season exists
- */
-function checkSeasonExists(season) {
-  if (season) return true;
-
-  const error = new Error("Season not found");
-
-  error.status = 404;
-  throw error;
-}
-
-/**
- * Updates the status of a Season.
- * If "save" is true, it will first save the changes to the Season.
- * @param {number} seasonId The ID of the season to update
- * @param {string} status The new status to set for the season
- * @param {boolean} savedWithErrors Whether the form was submitted with validation errors
- * @param {Object} [options={}] Optional season state updates
- * @param {boolean} [options.readyToPublish=null] Optionally provide a new readyToPublish value to set
- * @param {boolean} [options.informationSvcApproved=null] Optional Information Services approval value to set
- * @param {boolean} [options.reservationSvcApproved=null] Optional Reservation Services approval value to set
- * @param {Transaction} [options.transaction=null] Optional Sequelize transaction object for atomic operations
- * @returns {Promise<Season>} The updated season model
- */
-async function updateStatus(
-  seasonId,
-  status,
-  savedWithErrors,
-  {
-    readyToPublish = null,
-    informationSvcApproved = null,
-    reservationSvcApproved = null,
-    transaction = null,
-  } = {},
-) {
-  const season = await Season.findByPk(seasonId, { transaction });
-
-  checkSeasonExists(season);
-
-  // Update season status
-  season.status = status;
-
-  // Update the "savedWithErrors" flag
-  season.savedWithErrors = savedWithErrors;
-
-  // Update the "Ready to publish" flag if provided
-  if (readyToPublish !== null) {
-    season.readyToPublish = readyToPublish;
-  }
-
-  if (informationSvcApproved !== null) {
-    season.informationSvcApproved = informationSvcApproved;
-  }
-
-  if (reservationSvcApproved !== null) {
-    season.reservationSvcApproved = reservationSvcApproved;
-  }
-
-  // Update the updatedAt timestamp
-  season.updatedAt = new Date();
-
-  return season.save({
-    transaction,
-  });
-}
-
-/**
- * Returns the previous Season's dates for a given current Season.
- * @param {Season} currentSeason The current season object with operatingYear and publishableId
- * @param {Object} dateTypeWhere Optional where clause for filtering DateTypes
- * @returns {Array} Array with any DateRanges from the previous Season
- */
-async function getPreviousSeasonDates(currentSeason, dateTypeWhere = {}) {
-  try {
-    // @TODO: the previous season dates here are not the same as the /park endpoint
-    const prevSeason = await Season.findOne({
-      where: {
-        operatingYear: currentSeason.operatingYear - 1,
-        publishableId: currentSeason.publishableId,
-        seasonType: currentSeason.seasonType,
-      },
-      include: [
-        {
-          model: DateRange,
-          as: "dateRanges",
-          required: false,
-
-          include: [
-            {
-              model: DateType,
-              as: "dateType",
-              required: false,
-              attributes: ["id", "dateTypeNumber", "name"],
-
-              // Filter DateTypes by level
-              where: dateTypeWhere,
-            },
-          ],
-        },
-      ],
-    });
-
-    // If no previous season exists in the DB, return an empty array
-    if (!prevSeason) return [];
-
-    return prevSeason.dateRanges;
-  } catch (error) {
-    console.error("Error fetching previous season:", error);
-    throw error;
-  }
-}
-
-/**
- * Returns all DateRangeAnnuals for a given publishableId.
- * @param {number} publishableId The ID of the Publishable to get DateRange
- * @returns {Promise<Array>} An array of DateRangeAnnual models with their DateType
- */
-async function getDateRangeAnnuals(publishableId) {
-  if (!publishableId) return [];
-  return await DateRangeAnnual.findAll({
-    where: { publishableId },
-    attributes: ["id", "dateableId", "isDateRangeAnnual"],
-    include: [
-      {
-        model: DateType,
-        as: "dateType",
-        attributes: ["id", "dateTypeNumber", "name"],
-      },
-    ],
-  });
-}
-
-/**
- * Returns all GateDetails for a given publishableId.
- * @param {number} publishableId The ID of the Publishable to get GateDetail
- * @returns {Promise<Object|null>} GateDetail model, or null if not found
- */
-async function getGateDetail(publishableId) {
-  if (!publishableId) return null;
-  return await GateDetail.findOne({
-    where: { publishableId },
-    attributes: [
-      "id",
-      "hasGate",
-      "gateOpenTime",
-      "gateCloseTime",
-      "gateOpensAtDawn",
-      "gateClosesAtDusk",
-    ],
-  });
-}
-
-/**
- * Returns a query part for including change logs associated with a Season.
- * @returns {Object} Sequelize query part for fetching change logs
- */
-function changeLogsQueryPart() {
-  return {
-    model: SeasonChangeLog,
-    as: "changeLogs",
-    attributes: ["id", "notes", "createdAt"],
-    // Filter out empty notes
-    where: {
-      notes: {
-        [Op.ne]: "",
-      },
-    },
-    required: false,
-    order: [["createdAt", "DESC"]],
-    include: [
-      {
-        model: User,
-        as: "user",
-        attributes: ["id", "name"],
-      },
-    ],
-  };
-}
-
-/**
- * Returns a query part for including DateRanges associated with a Season.
- * @param {number} seasonId the ID of the DateRanges' Season
- * @returns {Object} Sequelize query part for fetching DateRanges
- */
-function dateRangesQueryPart(seasonId) {
-  return {
-    model: DateRange,
-    as: "dateRanges",
-    attributes: ["id", "startDate", "endDate", "dateTypeId", "dateableId"],
-    where: {
-      seasonId,
-    },
-    required: false,
-    order: [["startDate", "ASC"]], // @TODO: This doesn't work?
-    include: [
-      {
-        model: DateType,
-        as: "dateType",
-        attributes: ["id", "dateTypeNumber", "name"],
-      },
-    ],
-  };
-}
-
-/**
- * Returns a query part for including a Dateable and its DateRanges.
- * @param {number} seasonId the ID of the DateRanges' Season
- * @returns {Object} Sequelize query part for fetching Dateable and its DateRanges
- */
-function dateableAndDatesQueryPart(seasonId) {
-  return {
-    model: Dateable,
-    as: "dateable",
-    include: [dateRangesQueryPart(seasonId)],
-  };
-}
-
-/**
- * Returns a query part for including FeatureType details with a Feature.
- * @returns {Object} Sequelize query part for fetching FeatureType details
- */
-function featureTypeQueryPart() {
-  return {
-    model: FeatureType,
-    as: "featureType",
-    attributes: ["id", "name", "icon", "featureTypeNumber"],
-  };
-}
-
-// Common attributes for all Season queries
-const SEASON_ATTRIBUTES = [
-  "id",
-  "operatingYear",
-  "status",
-  "informationSvcApproved",
-  "reservationSvcApproved",
-  "readyToPublish",
-  "editable",
-  "publishableId",
-  "seasonType",
-  "savedWithErrors",
-];
-
-/**
- * Returns all Frontcountry Campground feature reservation dates for a park and operating year.
- * @param {Object} park Park model with features and parkAreas
- * @param {number} operatingYear Operating year for the Seasons
- * @returns {Promise<Array>} - Array of Frontcountry Campground feature reservation dates
- */
-async function getFrontcountryFeatureReservationDates(park, operatingYear) {
-  // Only fetch dates if the park has Winter fee dates or either Tier 1 or Tier 2 dates.
-  // This data is needed for Winter/Tier date validation.
-  if (!(park.hasWinterFeeDates || park.hasTier1Dates || park.hasTier2Dates))
-    return [];
-
-  const featurePublishableIds = park.features
-    // Filter out any park features without Publishable IDs
-    .filter((feature) => feature.publishableId)
-    .map((feature) => feature.publishableId);
-
-  const areaFeaturePublishableIds = park.parkAreas
-    // Filter out any park areas without Publishable IDs
-    .filter((parkArea) => parkArea.publishableId)
-    .map((parkArea) => parkArea.publishableId);
-
-  // Query the Season IDs for each publishable Feature in the Park,
-  // so we can look up their DateRanges
-  const featureSeasons = await Season.findAll({
-    attributes: ["id"],
-
-    where: {
-      operatingYear,
-
-      publishableId: {
-        [Op.in]: [...featurePublishableIds, ...areaFeaturePublishableIds],
-      },
-    },
-  });
-
-  const featureSeasonIds = featureSeasons.map((season) => season.id);
-
-  // Get all Frontcountry Campground Reservation DateRanges for these Seasons
-  return DateRange.findAll({
-    attributes: ["id", "startDate", "endDate", "dateTypeId", "seasonId"],
-    where: {
-      seasonId: {
-        [Op.in]: featureSeasonIds,
-      },
-
-      // Filter out blank date ranges (null startDate and endDate)
-      startDate: {
-        [Op.ne]: null,
-      },
-      endDate: {
-        [Op.ne]: null,
-      },
-    },
-
-    include: [
-      // Only include DateRanges with the Reservation DateType
-      {
-        model: DateType,
-        as: "dateType",
-        attributes: ["id"],
-        where: {
-          dateTypeNumber: DATE_TYPE.RESERVATION,
-        },
-        required: true,
-      },
-
-      // Only include DateRanges that are associated with a Frontcountry Campground feature
-      {
-        model: Dateable,
-        as: "dateable",
-        attributes: ["id"],
-        required: true,
-        include: [
-          {
-            model: Feature,
-            as: "feature",
-            required: true,
-            attributes: ["id"],
-            include: [
-              {
-                model: FeatureType,
-                as: "featureType",
-                attributes: ["id"],
-                where: {
-                  featureTypeNumber: FEATURE_TYPE.FRONTCOUNTRY_CAMPGROUND,
-                },
-                required: true,
-              },
-            ],
-          },
-        ],
-      },
-    ],
-  });
-}
-
-/**
- * Returns Tier 1, Tier 2, and Winter fee dates for a Park operating year.
- * @param {Object} park Park model with hasTier1Dates, hasTier2Dates, and publishableId
- * @param {number} operatingYear Operating year for the Seasons
- * @returns {Promise<Object>} - Object with parkTier1Dates, parkTier2Dates, and parkWinterDates arrays
- */
-async function getParkDates(park, operatingYear) {
-  // Query both regular and winter seasons to get the Tier 1, Tier 2, and Winter fee dates
-  // for this park and year
-  const seasons = await Season.findAll({
-    where: {
-      publishableId: park.publishableId,
-      operatingYear,
-      seasonType: [SEASON_TYPE.REGULAR, SEASON_TYPE.WINTER],
-    },
-
-    include: [
-      {
-        model: DateRange,
-        as: "dateRanges",
-        required: false,
-
-        include: [
-          {
-            model: DateType,
-            as: "dateType",
-            attributes: ["id", "dateTypeNumber", "name"],
-          },
-        ],
-      },
-    ],
-  });
-
-  // Merge all dateRanges from both seasons
-  const allRanges = seasons.flatMap((season) =>
-    (season.dateRanges || []).filter(
-      (range) => range.startDate && range.endDate,
-    ),
-  );
-
-  // Group by dateTypeNumber
-  const datesByStrapiId = _.groupBy(
-    allRanges,
-    (range) => range.dateType?.dateTypeNumber,
-  );
-
-  // Use constants for Tier 1, Tier 2, and Winter fee
-  const tier1Dates = datesByStrapiId[DATE_TYPE.TIER_1] || [];
-  const tier2Dates = datesByStrapiId[DATE_TYPE.TIER_2] || [];
-  const parkWinterDates = datesByStrapiId[DATE_TYPE.WINTER_FEE] || [];
-
-  // Only include tier dates if park supports them
-  const parkTier1Dates = park.hasTier1Dates ? tier1Dates : [];
-  const parkTier2Dates = park.hasTier2Dates ? tier2Dates : [];
-
-  return {
-    parkTier1Dates,
-    parkTier2Dates,
-    parkWinterDates,
-  };
-}
-
-/**
- * Returns the winter season for a park if it has winter fee dates enabled.
- * @param {Object} park Park model with hasWinterFeeDates and publishableId
- * @param {number} operatingYear Operating year for the Seasons
- * @returns {Promise<Object|null>} The winter season object with all related data, or null
- */
-async function getWinterSeason(park, operatingYear) {
-  if (!park.hasWinterFeeDates) {
-    return null;
-  }
-
-  // Find the winter season ID
-  const winterSeasonLookup = await Season.findOne({
-    attributes: ["id"],
-    where: {
-      publishableId: park.publishableId,
-      operatingYear,
-      seasonType: SEASON_TYPE.WINTER,
-    },
-  });
-
-  if (!winterSeasonLookup) {
-    return null;
-  }
-
-  const winterSeason = await Season.findByPk(winterSeasonLookup.id, {
-    attributes: SEASON_ATTRIBUTES,
-    include: [
-      {
-        model: Park,
-        as: "park",
-        include: [
-          // Park-level dates for this winter season
-          dateableAndDatesQueryPart(winterSeasonLookup.id),
-        ],
-      },
-
-      changeLogsQueryPart(),
-    ],
-  });
-
-  if (!winterSeason) {
-    return null;
-  }
-
-  // Get DateRangeAnnuals and GateDetail for winter season
-  const dateRangeAnnuals = await getDateRangeAnnuals(
-    winterSeason.publishableId,
-  );
-
-  return {
-    ...winterSeason.toJSON(),
-    dateRangeAnnuals,
-  };
-}
-
-/**
- * Saves season data (regular or winter season)
- * @param {Object} params Parameters for saving season data
- * @param {Season} params.season The season model instance
- * @param {Array} params.dateRanges Array of date ranges to save
- * @param {Array} params.dateRangeAnnuals Array of date range annuals to save
- * @param {Object|null} params.gateDetail Gate detail object (null for winter seasons)
- * @param {Object|null} params.oldGateDetail Existing gate detail object before save
- * @param {Array} params.deletedDateRangeIds Array of date range IDs to delete
- * @param {string} params.newStatus New status for the season
- * @param {boolean} params.informationSvcApproved Resolved Information Services team approval value
- * @param {boolean} params.reservationSvcApproved Resolved Reservation Services team approval value
- * @param {boolean|null} params.newReadyToPublish New readyToPublish value
- * @param {string} params.notes Notes for the change log
- * @param {boolean} params.savedWithErrors Whether the form was submitted with validation errors
- * @param {number} params.userId User ID making the changes
- * @param {Transaction} params.transaction Database transaction
- * @param {boolean} params.isWinterSeason Whether this is a winter season
- * @returns {Promise<void>}
- */
-async function saveSeasonData({
-  season,
-  dateRanges,
-  dateRangeAnnuals,
-  gateDetail,
-  oldGateDetail,
-  deletedDateRangeIds,
-  newStatus,
-  informationSvcApproved,
-  reservationSvcApproved,
-  newReadyToPublish,
-  notes,
-  savedWithErrors,
-  userId,
-  transaction,
-  isWinterSeason = false,
-}) {
-  // Calculate the actual new readyToPublish value
-  const actualNewReadyToPublish = newReadyToPublish ?? season.readyToPublish;
-
-  // Get the Winter Fee DateType's database ID
-  const winterFeeDateType = await DateType.findOne({
-    attributes: ["id"],
-    where: {
-      dateTypeNumber: DATE_TYPE.WINTER_FEE,
-    },
-    transaction,
-  });
-
-  if (!winterFeeDateType) {
-    throw new Error("Required DateType WINTER_FEE not found in the database.");
-  }
-
-  const winterFeeDateTypeId = winterFeeDateType.id;
-
-  // Filter date ranges based on season type
-  // Winter seasons should only have Winter fee dates
-  // Regular seasons should NOT have Winter fee dates
-  const filteredDateRanges = (dateRanges || []).filter((dateRange) => {
-    if (!dateRange.dateTypeId) return true;
-
-    if (isWinterSeason) {
-      return dateRange.dateTypeId === winterFeeDateTypeId;
-    }
-
-    return dateRange.dateTypeId !== winterFeeDateTypeId;
-  });
-
-  // dateRangeAnnuals
-  const dateRangeAnnualsToSave = (dateRangeAnnuals || []).map(
-    (dateRangeAnnual) => ({
-      id: dateRangeAnnual.id,
-      dateTypeId: dateRangeAnnual.dateType?.id,
-      publishableId: season.publishableId,
-      dateableId: dateRangeAnnual.dateableId,
-      isDateRangeAnnual: dateRangeAnnual.isDateRangeAnnual,
-    }),
-  );
-
-  // Upsert dateRangeAnnuals
-  const saveDateRangeAnnuals = DateRangeAnnual.bulkCreate(
-    dateRangeAnnualsToSave,
-    {
-      updateOnDuplicate: ["isDateRangeAnnual", "updatedAt"],
-      transaction,
-    },
-  );
-
-  // Handle gateDetail for regular seasons only
-  let saveGateDetail = Promise.resolve();
-  let gateDetailToSave = null;
-
-  if (!isWinterSeason && gateDetail) {
-    gateDetailToSave = {
-      ...gateDetail,
-      publishableId: season.publishableId,
-    };
-
-    saveGateDetail = GateDetail.upsert(gateDetailToSave, {
-      transaction,
-    });
-  }
-
-  // Create season change log with the notes
-  const seasonChangeLog = await SeasonChangeLog.create(
-    {
-      seasonId: season.id,
-      userId,
-      notes,
-      statusOldValue: season.status,
-      statusNewValue: newStatus,
-      readyToPublishOldValue: season.readyToPublish,
-      readyToPublishNewValue: actualNewReadyToPublish,
-      gateDetailOldValue: oldGateDetail,
-      gateDetailNewValue: gateDetailToSave,
-    },
-    { transaction },
-  );
-
-  // Update the season object with the new status and readyToPublish values
-  const saveSeason = updateStatus(season.id, newStatus, savedWithErrors, {
-    readyToPublish: newReadyToPublish,
-    informationSvcApproved,
-    reservationSvcApproved,
-    transaction,
-  });
-
-  // Create date change logs for updated dateRanges
-  const existingDateIds = filteredDateRanges
-    .filter((date) => date.id)
-    .map((date) => date.id);
-
-  let createChangeLogs = Promise.resolve();
-
-  if (existingDateIds.length > 0) {
-    const existingDateRows = await DateRange.findAll({
-      where: {
-        id: {
-          [Op.in]: existingDateIds,
-        },
-      },
-      transaction,
-    });
-
-    const datesToUpdateById = _.keyBy(filteredDateRanges, "id");
-    const changeLogsToCreate = existingDateRows.map((oldDateRange) => {
-      const newDateRange = datesToUpdateById[oldDateRange.id];
-
-      return {
-        dateRangeId: oldDateRange.id,
-        seasonChangeLogId: seasonChangeLog.id,
-        startDateOldValue: oldDateRange.startDate,
-        startDateNewValue: newDateRange.startDate,
-        endDateOldValue: oldDateRange.endDate,
-        endDateNewValue: newDateRange.endDate,
-      };
-    });
-
-    createChangeLogs = DateChangeLog.bulkCreate(changeLogsToCreate, {
-      transaction,
-    });
-  }
-
-  // Update or create dateRanges
-  let updateDates = Promise.resolve();
-
-  if (filteredDateRanges.length > 0) {
-    updateDates = DateRange.bulkCreate(filteredDateRanges, {
-      updateOnDuplicate: ["startDate", "endDate", "updatedAt"],
-      transaction,
-    });
-  }
-
-  // Delete dateRanges removed by the user
-  let deleteDates = Promise.resolve();
-
-  if (deletedDateRangeIds.length > 0) {
-    deleteDates = DateRange.destroy({
-      where: {
-        id: {
-          [Op.in]: deletedDateRangeIds,
-        },
-      },
-      transaction,
-    });
-  }
-
-  await Promise.all([
-    saveSeason,
-    updateDates,
-    createChangeLogs,
-    deleteDates,
-    saveDateRangeAnnuals,
-    saveGateDetail,
-  ]);
-}
 
 // Get all form data and DateRanges for a Feature Season
 router.get(
